Show error message when trying to delete admin account

+++ app/Http/Controllers/UserController.php
@@ -44,6 +44,10 @@ class UserController extends Controller
             'user_id.*' => 'The selected user is invalid or does not exist.'
         ]);
 
+        if (User::query()->find($request->input('user_id'))->admin) {
+            return response()->json(['success' => false, 'message' => 'You cannot delete an admin account.'],403);
+        }
+
         User::query()->where('id', '=', $request->input('user_id'))->where('admin', '=', 'false')->delete();
 
         return response()->json(['success' => true, 'user_id' => $request->input('user_id')]);
